feat(routing): Support session based routing
The session id field in a message buffer is now used
directly for routing.


Routing table entry changes to pick up subid


Fix unit test for rmr_call

It was not properly setting message type and now that RMr
ensures that invalid message type is set by default on a newly
created message this was causing unit test to fail.


Add application level tests

Added round robin and multi group application level
test scripts.


NNG based sub-id support added


Correct bug in app test script


Nano sub-id changes and unit test updates


Change nil-sub_id key to high order key

If there is no sub-id, then the key is based only on the
message type, but to allow for a sub-id == 0 the key
when there is no subscription id must be set to
0xffffffff00000000 + msg type.

New version for deb is 1.0.19

+++ src/common/include/rmr_symtab.h
@@ -28,6 +28,8 @@
 #ifndef _rmr_symtab_h
 #define _rmr_symtab_h
 
+#include <netdb.h>
+
 /* --------- symtab ---------------- */
 #define UT_FL_NOCOPY 0x00          /* use user pointer */
 #define UT_FL_COPY 0x01            /* make a copy of the string data */
@@ -39,12 +41,12 @@ extern void rmr_sym_clear( void *s );
 extern void rmr_sym_dump( void *s );
 extern void *rmr_sym_alloc( int size );
 extern void rmr_sym_del( void *s, const char *name, unsigned int class );
-extern void *rmr_sym_ndel( void *vtable, int key );
+extern void *rmr_sym_ndel( void *vtable, uint64_t key );
 extern void rmr_sym_free( void *vtable );
 extern void *rmr_sym_get( void *s,  const char *name, unsigned int class );
 extern int rmr_sym_put( void *s,  const char *name, unsigned int class, void *val );
-extern int rmr_sym_map( void *s,  unsigned int key, void *val );
-extern void *rmr_sym_pull(  void *vtable, int key );
+extern int rmr_sym_map( void *s,  uint64_t key, void *val );
+extern void *rmr_sym_pull(  void *vtable, uint64_t key );
 extern void rmr_sym_stats( void *s, int level );
 extern void rmr_sym_foreach_class( void *vst, unsigned int class, void (* user_fun)( void*, void*, const char*, void*, void* ), void *user_data );
